tried refactoring, broke things
