+++ soh/src/code/audio_load.c
@@ -5,7 +5,7 @@
 #include <libultraship/libultra.h>
 #include "global.h"
 #include "soh/OTRGlobals.h"
-#include "soh/Enhancements/sfx-editor/SfxEditor.h"
+#include "soh/Enhancements/audio/AudioEditor.h"
 
 #define MK_ASYNC_MSG(retData, tableType, id, status) (((retData) << 24) | ((tableType) << 16) | ((id) << 8) | (status))
 #define ASYNC_TBLTYPE(v) ((u8)(v >> 16))
